move OutputSorter to composer, changed protocol version to 3.0t
